Refactor: Move main file into multiple files
