+++ Windows/Gopher/Gopher/CXBOXController.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <Windows.h>
+#include <Xinput.h>
+
+class CXBOXController
+{
+private:
+	XINPUT_STATE _controllerState;
+	int _controllerNum;
+public:
+	CXBOXController(int playerNumber);
+	XINPUT_STATE GetState();
+	bool IsConnected();
+};
