+++ Windows/Gopher/Gopher/Gopher.vcxproj
@@ -76,7 +76,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="CXBOXController.cpp" />
+    <ClCompile Include="Gopher.cpp" />
     <ClCompile Include="main.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="ButtonState.h" />
+    <ClInclude Include="CXBOXController.h" />
+    <ClInclude Include="Gopher.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
